# src/openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCJars Versions API",
    "version": "1.0.0",
    "contact": {
      "email": "rjansengd@gmail.com",
      "name": "GitHub",
      "url": "https://github.com/mcjars/versions-worker"
    }
  },
  "paths": {
...
    "/api/v2/stats/{type}/{version}": {
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "stats": {
                      "type": "object",
                      "properties": {
                        "builds": {
                          "type": "integer"
                        },
                        "hashes": {
                          "type": "integer"
                        },
                        "size": {
                          "type": "object",
                          "properties": {
                            "total": {
                              "type": "object",
                              "properties": {
                                "jar": {
                                  "type": "integer"
                                },
                                "zip": {
                                  "type": "integer"
                                }
                              },
                              "required": [
                                "jar",
                                "zip"
                              ]
                            },
                            "average": {
                              "type": "object",
                              "properties": {
                                "jar": {
                                  "type": "integer"
                                },
                                "zip": {
                                  "type": "integer"
                                }
                              },
                              "required": [
                                "jar",
                                "zip"
                              ]
                            }
                          },
                          "required": [
                            "total",
                            "average"
                          ]
                        }
                      },
                      "required": [
                        "builds",
                        "hashes",
                        "size"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "stats"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Version not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "description": "Get Stats of all Builds for a Version of a Project",
        "parameters": [
          {
            "description": "Project Type",
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/types"
            }
          },
          {
            "description": "Version",
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1.17.1"
            }
          }
        ]
      }
    },
    "/api/v2/stats/{type}/{version}/history/{year}/{month}": {
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "stats": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "builds": {
                            "type": "integer"
                          },
                          "hashes": {
                            "type": "integer"
                          },
                          "size": {
                            "type": "object",
                            "properties": {
                              "total": {
                                "type": "object",
                                "properties": {
                                  "jar": {
                                    "type": "integer"
                                  },
                                  "zip": {
                                    "type": "integer"
                                  }
                                },
                                "required": [
                                  "jar",
                                  "zip"
                                ]
                              },
                              "average": {
                                "type": "object",
                                "properties": {
                                  "jar": {
                                    "type": "integer"
                                  },
                                  "zip": {
                                    "type": "integer"
                                  }
                                },
                                "required": [
                                  "jar",
                                  "zip"
                                ]
                              }
                            },
                            "required": [
                              "total",
                              "average"
                            ]
                          }
                        },
                        "required": [
                          "builds",
                          "hashes",
                          "size"
                        ]
                      }
                    }
                  },
                  "required": [
                    "success",
                    "stats"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Version not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "description": "Get Stats of all Builds for a Version of a Project",
        "parameters": [
          {
            "description": "Project Type",
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/types"
            }
          },
          {
            "description": "Version",
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1.17.1"
            }
          },
          {
            "description": "Year",
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 2024,
              "min": 2024
            }
          },
          {
            "description": "Month",
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1,
              "min": 1,
              "max": 12
            }
          }
        ]
      }
    },
    "/api/v2/stats/{type}": {
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "stats": {
                      "type": "object",
                      "properties": {
                        "builds": {
                          "type": "integer"
                        },
                        "hashes": {
                          "type": "integer"
                        },
                        "size": {
                          "type": "object",
                          "properties": {
                            "total": {
                              "type": "object",
                              "properties": {
                                "jar": {
                                  "type": "integer"
                                },
                                "zip": {
                                  "type": "integer"
                                }
                              },
                              "required": [
                                "jar",
                                "zip"
                              ]
                            },
                            "average": {
                              "type": "object",
                              "properties": {
                                "jar": {
                                  "type": "integer"
                                },
                                "zip": {
                                  "type": "integer"
                                }
                              },
                              "required": [
                                "jar",
                                "zip"
                              ]
                            }
                          },
                          "required": [
                            "total",
                            "average"
                          ]
                        }
                      },
                      "required": [
                        "builds",
                        "hashes",
                        "size"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "stats"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "description": "Get Stats of all Builds for a Project",
        "parameters": [
          {
            "description": "Project Type",
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/types"
            }
          }
        ]
      }
    },
    "/api/v2/stats/{type}/history/{year}/{month}": {
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "stats": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "builds": {
                            "type": "integer"
                          },
                          "hashes": {
                            "type": "integer"
                          },
                          "size": {
                            "type": "object",
                            "properties": {
                              "total": {
                                "type": "object",
                                "properties": {
                                  "jar": {
                                    "type": "integer"
                                  },
                                  "zip": {
                                    "type": "integer"
                                  }
                                },
                                "required": [
                                  "jar",
                                  "zip"
                                ]
                              },
                              "average": {
                                "type": "object",
                                "properties": {
                                  "jar": {
                                    "type": "integer"
                                  },
                                  "zip": {
                                    "type": "integer"
                                  }
                                },
                                "required": [
                                  "jar",
                                  "zip"
                                ]
                              }
                            },
                            "required": [
                              "total",
                              "average"
                            ]
                          }
                        },
                        "required": [
                          "builds",
                          "hashes",
                          "size"
                        ]
                      }
                    }
                  },
                  "required": [
                    "success",
                    "stats"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "description": "Get Stats of all Builds for a Project",
        "parameters": [
          {
            "description": "Project Type",
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/types"
            }
          },
          {
            "description": "Year",
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 2024,
              "min": 2024
            }
          },
          {
            "description": "Month",
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1,
              "min": 1,
              "max": 12
            }
          }
        ]
      }
    },
...
    "/api/v2/stats/version/{version}": {
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "stats": {
                      "type": "object",
                      "properties": {
                        "builds": {
                          "type": "integer"
                        },
                        "hashes": {
                          "type": "integer"
                        },
                        "size": {
                          "type": "object",
                          "properties": {
                            "total": {
                              "type": "object",
                              "properties": {
                                "jar": {
                                  "type": "integer"
                                },
                                "zip": {
                                  "type": "integer"
                                }
                              },
                              "required": [
                                "jar",
                                "zip"
                              ]
                            },
                            "average": {
                              "type": "object",
                              "properties": {
                                "jar": {
                                  "type": "integer"
                                },
                                "zip": {
                                  "type": "integer"
                                }
                              },
                              "required": [
                                "jar",
                                "zip"
                              ]
                            }
                          },
                          "required": [
                            "total",
                            "average"
                          ]
                        }
                      },
                      "required": [
                        "builds",
                        "hashes",
                        "size"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "stats"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Version not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "description": "Get Stats of all Builds for all Projects for a minecraft version",
        "parameters": [
          {
            "description": "Minecraft Version",
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1.17.1"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "types": {
        "type": "string",
        "enum": [
          "VANILLA",
          "PAPER",
          "PUFFERFISH",
          "FOLIA",
          "PURPUR",
          "WATERFALL",
          "VELOCITY",
          "FABRIC",
          "BUNGEECORD",
          "QUILT",
          "FORGE",
          "NEOFORGE",
          "MOHIST",
          "ARCLIGHT",
          "SPONGE",
          "LEAVES"
        ]
      }
    },
...
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "scheme": "token"
      }
    },
...
  },
...
  "servers": [
    {
      "url": "https://versions.mcjars.app"
    }
  ]
}